PMM-6091 Rename UpdateInfoBox to InfoBox
